convert toolkit_page to concern
