+++ app/models/concerns/editable_page.rb
@@ -1,0 +1,32 @@
+# A module for editable pages
+#   required fields:
+#     - name
+#     - markdown
+#     - last_user_id
+module EditablePage
+  extend ActiveSupport::Concern
+
+  included do
+    belongs_to :last_user, foreign_key: "last_user_id", class_name: 'User'
+    validates :name, uniqueness: true, presence: true
+    before_validation :modify_name
+    before_create :set_markdown_to_be_blank_string_if_nil
+  end
+
+  class_methods do
+    def pagify_name(name)
+      name.downcase.tr(' ', '_')
+    end
+  end
+
+  private
+
+  def set_markdown_to_be_blank_string_if_nil
+    self.markdown = "" if markdown.nil?
+  end
+
+  def modify_name
+    return if persisted? || name.nil?
+    self.name = self.class.pagify_name(name)
+  end
+end
