+++ spec/models/toolkit_page_spec.rb
@@ -1,11 +1,18 @@
 require 'rails_helper'
 
 describe ToolkitPage, type: :model do
+  it { should belong_to(:last_user) }
+
+  describe 'pagify_name' do
+    it 'converts to lowercase and replaces spaces with underscores' do
+      expect(ToolkitPage.pagify_name('My Name')).to eq 'my_name'
+    end
+  end
+
   describe 'validations' do
     subject { ToolkitPage.new(name: "page") }
     it { should validate_uniqueness_of(:name).case_insensitive }
     it { should validate_presence_of(:name) }
-    it { should belong_to(:last_user) }
   end
 
   describe 'set_markdown_to_be_blank_string_if_null' do
